# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MobiManager API
  description: Complete API documentation for MobiManager - a mobile shop management SaaS platform
  version: 1.0.0
  contact:
    name: MobiManager Support
servers:
  - url: http://localhost:3000
    description: Local development server
security:
  - cookieAuth: []
paths:
...
  /api/admin/ai/festival-offers:
    post:
      summary: Generate Festival Offers
      tags: [AI]
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Festival offers
  /api/admin/ai/slow-stock:
    post:
      summary: Get Slow Stock Suggestions
      tags: [AI]
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Slow stock suggestions
  /api/admin/ai/monthly-strategy:
    post:
      summary: Generate Monthly Strategy
      tags: [AI]
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Monthly strategy
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: admin_token